+++ azai/literature/__init__.py
@@ -1,0 +1,1 @@
+"""Lightweight local literature retrieval utilities."""